folder restructure, added extract_education
