--- utilties/job_description_parser.py
@@ -1,6 +1,0 @@
-class JobDescriptionParser:
-    def __init__(self, job_description_path, debug=False):
-        self.job_description = job_description_path
-
-    def parse(self):
-        pass
