+++ utilties/job_description_parser/job_description_parser.py
@@ -1,0 +1,6 @@
+class JobDescriptionParser:
+    def __init__(self, job_description):
+        self.job_description = job_description
+
+    def parse(self):
+        pass
